docs(filter-select): update FilterSelect docs to use CustomCanvas
